+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Debug card to retrieve diagnostic information
 - UART for programming, configuration and debugging instead of USB CDC
 - Card readout, formatting via shell
+- Cyclic card formatting, reading previous run punches
 
 ### Changed
 
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Use of external EEPROM
+- Clearing punches in the Android app
 
 ## [2.1.0] - 2025.11.30